Implement a new outlining algorithm.
The previous algorithm to detect when to start and stop outlining was
complicated and required the existence of unmappable blocks in the trace. This
put an an extra burden on the software tracer which would have to simulate those
unmappable blocks, a sometimes difficult task.

The new algorithm is much simpler and works without unmappable blocks being
present. It works as follows:

1) We detect a function that requires outlining. We note the successor block
   `succ` of the current block by looking at the current block's branch
   instruction. When we see this block outlining is finished.
2) Continue processing blocks. For each block check the following:
   a) is this block an entry block to the same function in which we started
      outlining? If yes, then we have recursion: increment a counter.
   b) is this block an exit block (i.e. it contains a return instruction) of
      the same function in which we started outlining? If yes, we are returning
      from a recursive call: decrement the counter.
   c) is the recursion counter zero and this block is the successor block
      `succ`? If yes, then stop outlining.
   d) if none of the above is true, skip this block entirely.