Add `DeoptSafepoint` instruction
This solves the issue where deserialisation and handling of calls from
within the yk runtime were special-cased to handle LLVM's stack-map
intrinsic. This was brittle and created complex nested pattern matching.

Now, calls to LLVM's stack-map intrinsics are serialised to disk as
`DeoptSafepoint` instructions, where the live variables are easily
retrievable, and useless duplication (such as the function name) is
removed.