Make the constant pool a set.
We previously used a `Vec` (well, a `TiVec`) in a manner that was
effectively a set (e.g. using `position`) that was both slow (O(n)
insertions!) and obfuscatory.

This commit explicitly uses a set to represent the constant pool. But we
don't use any old set type -- we use `IndexSet`. That gives us set-speed
for insertion but vec-speed for lookups, because one can look things up
by index.