Simplify `CallInst` so it knows how many arguments it takes.
This isn't useful right now, but will shortly allow us to fold in
`VACall` too. In order to achieve this, we push all arguments straight
on to `m.extra_args`.

We might one day want to slightly optimise this by breaking out
`Call0` ("call with no arguments"), `Call1` ("call with 1 argument"),
`CallN` ("call with 'n' arguments") and so on, but at the moment this is
probably a very minor optimisation at most.