Overhaul PtrAdd to allow dynamic offsetting.
Yk PtrAdd instructions come from LLVM getelementptr (GEP) instructions,
which implement pointer arithmetic using field and element indices (of
struct and arrays).

We lower a LLVM GEPs to Yk AOT PtrAdd instructions which implement the
same semantics, but instead of using field and element indices, they use
a byte offset. This is simpler for us, as it means we don't have to walk
through aggregate types at runtime to find the byte offsets (which is
all the codegen really needs to know).

The current design of PtrAdd adds a constant value to a pointer, but
some programs require dynamic (computed at runtime) offsetting.

This change adapts the AOT PtrAdd opcode to allow dynamic offsetting.

Every getelementpointer instruction can be reduced to:
 - a constant offset
 - zero or more dynamic offsets of the form `elem_count * elem_size`,
   where `elem_count` may be dynamic.

So this is exactly what we stash into our AOT PtrAdd instruction. When
encountering the revised PtrAdd, the trace builder will emit Add and Mul
instructions in order to compute the dynamic offset at runtime.

Note that the JIT IR PtrAdd accepts a simple operand argument.

After this change simple_interp_loop{1,2} work with the new codegen! A
little bit of a milestone :)