Turn an internal-mistake `panic` into a `debug_assert`.
It feels more in keeping with other code -- and will be very marginally
faster -- to turn `push_and_make_operand` into a function with a
precondition enforced in debug mode with `debug_assert` than a run-time
`panic`. If we do make this mistake, we've got something pretty wrong
elsewhere in tracebuilder.