Add a tentative optimisation module and example optimisation.
The aim of this commit isn't to turn the JIT compiler into a fully
optimising compiler, but to explore what we need to do to make it
practical to write optimisations. To do that, I've implemented a (very)
simple dead-code elimination pass that (to keep things simple for now)
only really works with `Add` instructions. This does work for the one
test I've implemented, but it's (ahem) not exactly production quality!

This has ended up being a more valuable exercise than I expected because
it has immediately highlighted several factors:

  1. Instructions need to be `Clone` and `PartialEq`.

  2. For testing purposes (but not non-testing!) we need some sort of
     instruction that consumes an operand and tells (a.k.a. "lies") to
     the optimiser that the instruction has a side effect. One way of
     thinking of this is that it's equivalent to a "print instruction
     that doesn't print". I've tentatively called it `Consume` but am
     open to other naming suggestions.

  3. We need a way of writing human-friendly syntax for test inputs (we
     have a way -- which I haven't used -- of using human-friendly
     syntax for test outputs, but that's only half the problem). The
     "solution" to this is semi-obvious, but it's not a trivial amount
     of work.

  4. It's probably going to be better adding functions to instructions
     to express things like "does this instruction produce a side
     effect" rather than encoding that knowledge in every pass via
     (tediously long) `match` statements.

  5. We might want to think about things like adding a "rewrite operand"
     function/visitor to `Inst` rather than encoding it in each pass.
     I'm less sure about that, though.

I doubt these are the only things that we'll come across as we implement
optimisations, but it seems like a useful starting point. So this commit
is both, to some extent, a request for comments, but also "it's probably
worth thinking about merging this as it can be a very early guide as to
how we get the system ready for more powerful optimisations".