Make test run much less long.
This test was taking about 10s to run, which is ludicrous, given that it
only tests a minor property. This change reduces its running time to 1s,
which seems much more reasonable (and still, at least when I quickly
prodded it, still tests the things we want it to test).