Review human-readable JIT IR, in preparation for having it parsable.
This change hopefully makes the IR more consistent and concise. If
users/developers are going to be writing textual IR by hand, we may as
well make is easier for them.

 - Make all names lower case, using underscores where it feels right.
   e.g. `PtrAdd` -> `ptr_add`, `LookupGlobal` -> `lookup_global`.

 - Shorten some names too, notably:
   - Use LLVM short predicate names, because people know them. e.g.
     `eq`, `sgt` and `ule` (equal, signed-greater-than,
     unsigned-less-than-or-equal).
   - `IndirectCall` -> `icall`.
   - `LoadTraceInput` -> `load_ti`
   - `SignExtend` -> `sext`
   - `ZeroExtend` -> `zext`
   - `TraceLoopStart` -> `tloop_start` (and updated debug comments and
     dynasm label names for consistency)

 - Format `icmp` instructions as though each predicate is its own
   instruction. Later we should change the underlying data structures.

This mostly only touches the JIT IR for now. In one place a construct
was shared among both IRs, so that changed for both.

I'll do the same review for the AOT IR later.

While here, also document the `display()` idiom.

Reformat icmp.