Be both more flexible and less forgiving when JIT IR parsing.
This commit makes it possible to have JIT IR inputs which do not have
operand assignment [0, 1, 2, 3, ...] in order: operands can now be
defined in any order and the parser will map them appropriately. In
other words, the textual IR maps to IR that is equivalent modulo
alpha-renaming.

As part of this, the parser now catches some obvious mistakes
(referencing undefined operands, redefining operands).