Let fm do the hard work when pattern matching fails.
I think this code predates https://github.com/softdevteam/fm/pull/39 in
fm: fm now prints out the disassembly code, and then where the literal
and pattern didn't match. So by printing out the assembly twice (as well
as the shouty "!!!") we're just making it harder to work out what the
problem was.