Parse JIT IR for guards.
There is a design choice here: what `GuardInfo` should such a guard
have? At least for the time being my answer is "completely empty
`GuardInfo`". This might not always be what we want, at which point
we'll have to think about how we can add extra `GuardInfo` in a parsable
sense or via some other mechanism. But this is still better than
nothing.