Implement a tiny, but more realistic, optimisation pass.
`lvn.rs` was a very quick sketch of an optimisation pass, but it was
neither complete nor fully correct. Still, it did help me get an idea of
where we want to go.

This commit adds another tiny -- but less incorrect! -- optimisation
pass which does one single optimisation: it translates "mul X" to "shl
X" if X is a power of 2. This is often called "strength reduction",
though that is often intended as part of loop-based calculations.
RPython calls this, and similar, optimisations "rewrites" but that feels
too generic to me: really this is an instance of a very-local
very-simple optimisation.

I was hoping to get from this a better sense of how easy it is to
manipulate JIT IR. I got that, but I also ended up with a better
understanding of where our JIT IR seems not to follow LLVM IR, in a way
that's probably not a good idea for us longer term. Still, this commit
is useful, because it's a concrete bit of code for us to modify as we go
forwards (or not...).