Don't add an assignment instruction after every pointer add.
Before this commit we added an assignment instruction after every
pointer add to help map the AOT IR to the JIT IR. This commit removes
that assignment instruction, instead linking the AOT IR to the last
instruction in the JIT IR pointer add chain.

That also means that we no longer need `AssignInst` in the JIT IR at
all.