Skip duplicate control point blocks.
When processing a trace generated by the hardware tracer, trace_builder will
encounter two control point blocks. We must only process one of them. The
simplest way to do this that is compatible with the software tracer is to
start outlining after processing the first block by setting the outlining
target to the blocks successor.