Optimise `Mul`s without shuffling instruction indices.
The motivating part of this commit is to optimise `Mul` when it has at
least one constant:

  1. `x * 0` -> `0`
  2. `x * 1` -> `x`
  3. `x * y` -> `x << ...` when `y` is an appropriate constant

The first and second optimisation implicitly delete instructions, but we
don't want to actually remove instructions as we would have to shuffle
the backing vector.

Instead we introduce two new "virtual" instructions: `ProxyConst` and
`ProxyInst`. These can be thought of assignments (allowing, roughly, `%0
= 1i8` and `%3 = %2` respectively) but they are invisible to users of
the JIT IR.

To make this happen, a `PackedOperand` can no longer be directly
converted to an `Operand`: we have to account for proxy instructions.
Imagine we had JIT IR that, internally, looked like this:

```
%0 = 0i8 ; ProxyConst
%1 = add ...
%2 = %1 ; ProxyInst
%3 = add %0, %2
```

Users of the JIT IR will perceive the IR as if it was actually:

```
%1 = add ...
%3 = add 0i8, %1
```

i.e. the `ProxyConst` and `ProxyInst` will be invisible. Note, though,
that the instruction IDs `%1` `%3` have maintained their positions,
which is how we can infer that there are proxies in the underlying
vector.

This is a somewhat disruptive change across the JIT IR as we have to
thread `Module` references in many places where they previously didn't
occur. I'm sure we can finesse this further, but this feels like a
reasonable stopping point.