Rename `jit_ir::Module::len` to `Module::last_inst_idx`.
`len` was always a bit ambiguous: the length of what? There are many
things stored in `jit_ir::Module`. It also turns out that we always used
this in the idiom `m.len() - 1`, so really we didn't want to know the
number of instructions, but the index of the most recently pushed
instruction.

This commit makes clearer that what one is getting is the index of the
most recently pushed instruction.