Remove `TiVec` from `Module::insts`.
Internally to `jit_ir/mod.rs` using `InstIdx` is only a minor win, but
having `TiVec` forces us to allow `From` `usize` to `InstIdx`, which is
a footgun for all users of this module. The only way to avoid this is to
stop using `TiVec`: since users of `insts` have to go via functions
which mandate `InstIdx`, we still force users of this module to use the
newtype we want them to.