Make the skipping iterator also produce instructions.
In all (three...) places that we currently use this iterator, we
immediately use the index returned by the iterator to lookup an
instruction. This commit tweaks things to also return a reference to the
instruction itself. This makes it easier for the user of the iterator to
be sure that they're getting a non-`Proxy*`/`Tombstone` instruction.