Properly handle `Proxy*` objects.
I slightly underthought the introduction of `Proxy` objects, which
caused me problems when trying to extend the dead code elimination.

In essence, what I didn't fully consider is that `Proxy*` objects change
the IR "in memory" from the "IR we nearly always want to see".

For example if we IR in memory stored along these lines:

```text
%0: ...
%1: 1i8
%2: add %0, %1
```

Then instruction `%1` is a `ProxyConst`. We very rarely want to view the
IR as it is directly stored in memory. Instead, we nearly always want to
pretend that the IR really looks like:

```text
%0: ...
%2: add %0, 1i8
```

i.e. we want to pretend that the `%1` doesn't exist and that the
reference to `%1` is actually just a constant.

However, we didn't quite handle this correctly before. This commit fixes
two things:

  1. Most of the time when we ask for an `Inst` we don't want a `Proxy*`
     object. So there are now two functions for getting an `Inst`:
     `inst_no_proxies` panics if you try to read a `Proxy*`. You nearly
     always want to use this method! In those rare cases where you don't
     (and you really, really need to know that you don't!) there is
     `inst_all`. Use the latter with care!

  2. We can't ever store `Operand`s directly because an important part
     of "deproxification" is `PackedOperand::unpack`. Fortunately we
     only stored `Operand`s in `Module::args`, so switching those to
     `PackedOperand`s wasn't too difficult.

I'm sure this won't turn out to be the last time we encounter this
issue, but I am hopeful that at the very least this commit nudges us
quite a long way in the right direction.