Largely complete dead code elimination.
This commit (largely) fleshes out dead code elimination (modulo the
`Select` instruction, which I haven't seen in practise yet). The key
insight (which took me embarrassingly long to realise) is that it needs
to operate on `PackedOperand`s, not deproxied `Inst`s.

I have also enabled this as a pass in the optimiser. In the future, the
codegen's register allocator should naturally perform dead code
elimination and we won't need to use this code outside of testing mode.
However, that future is not here yet, and in real code (e.g. yklua),
dead code elimination already removes enough stuff to have a small but
noticeable performance improvement (some poor quality experiments
suggest ~5-10%).