Push decoding of the first block into the iterator.
Previously, stopping HWT tracing required disassembling the first part
of the trace synchronously (i.e. blocking the main thread). Although
this wasn't a huge cost, it is noticeable in a profile. This commit
modifies the iterator so that this happens in the first `next` call,
causing this to happen on a thread.