Rework the `CodeGen` trait.
The basic aim here is to make a `CodeGen` persistent: this lowers the
costs of compiling multiple things and also allows a codegen to share
state (e.g. caching) if it wants to.

In essence, we instantiate a `CodeGen` once and then repeatedly call
`codegen` on it:

```rust
pub(crate) trait CodeGen: Send + Sync {
  /// Generate code for the module `m`.
  fn codegen(&self, m: Module) -> Result<Arc<dyn CompiledTrace>, CompilationError>;
}
```

This idiom also matches how we deal with other similar parts of yk (e.g.
tracers): an "instantiate once, use many" idiom.

At them moment I have minimally hacked this into the x86 backend: we
could undoubtedly do a nicer, fuller, job, but this nudges us in the
right sort of direction.