Minimally deal with const ints of different sizes in the x86 backend.
This isn't really a complete solution, because at the moment we store
every constant integer in 64 bits worth of space, even if it's smaller.
If/when we use less space, we might get some fun bugs as we find it
breaks assumptions elsewhere, but it's difficult for us to deal with
that until we encounter it.