Avoid talking about "allocation" of SSA variables.
Like a surprising number of things in compilers, the "conventional"
terminology here is messy. Most obviously, we allocate registers, but
some of those allocated registers are not registers! Similarly, we might
allocate memory that's small enough that we can store the result in a
register, at which point we haven't allocated memory in the way that's
normally thought of.

To avoid compounding this, rename the struct `VarLocation`: there's no
perfect name here, but this hopefully at least breaks any supposed link
with "malloc".