Adjust test to shadowstack change.
We no longer adjust the shadowstack for external functions as there's no need to
do so. While this doesn't have a measurable performance gain, it does clean up
the IR quite a bit. So this

```
%1 = ptr_add shadowstack, 16
*shadowstack = %1
call external_func()
*shadowstack = %orig_shadowstack
```

turns into

```
call external_func()
```

The change broke a few tests which this commit fixes.