Execute the trace inside the AOT frame.
This change resets the stack back to the AOT frame before executing a trace.
Effectively, this pretends the control point never happened, placing the AOT
frame and trace frame right next to each other. This will allow us to easily
access the AOT frame from within the trace to read out live variables, which
in turn enables us to get rid of the trace inputs, in a future commit.