Fix sidetracing check in control point transition.
Previously, after compiling a side trace, we queried the root compiled
trace, but that's incorrect, because other threads may have done things
to it (e.g. started another side trace compiling). Not only is it
incorrect, but we don't even *need* to query it -- we have all the
information we need without doing so! In essence, we now say "if we're
compiling a side trace do X" rather than "if parent trace is Y do X".