Deserialise and display line-level debug info in the AOT IR.
This is the first step towards having C source code in gdb when you are
debugging a trace.

Unfortunately this requires some changes to `display()` for a few data
structures. Hopefully these changes aren't too annoying for those using
`display()` for debuging. At least the additional parameters are all
`Option`, so you can pass `None` if you don't have the required
information to hand.

I can already see the lines floating around a bit, but such is debug
info. It has to be better than nothing.