Allow type_() to work on global operands.
The type is always `Ty::Ptr` for globals.

This fixes a crash where when you try to `YKD_LOG_IR=-:aot` and you have
instructions that defer the type decision to operands which may be
globals, e.g. `phi` and `switch`. Test included.

(The test would have been better as a unit test, but manually
constructing an AOT module is painful, so I used an integration test)