Use a better heuristic for which registers to spill.
At some point every register allocator runs out of registers, and then
we have to decide which unfortunate register to spill and reuse.
Previously we semi-randomly picked a register (which, mostly, would end up
being R15) and spilt it.

This commit uses a common heuristic in linear scan register allocators:
spill the register whose value is last used furthest away. The idea here
is that we already know when a value is last used, and the later a value
is used, the more likely it is that it's not used soon. In other words,
this is a rough proxy for "find the value which is least likely to be
used soon and spill that".

Pleasingly, crude and simple though this is, it leads to noticeably less
spills overall: big_loop.lua speeds up by a consistent 5.5%.