Turn a `CompilationError` into an `unwrap`.
We're using `CompilationError` here as a way of not doing a check in
ykllvm, but if we ever hit this, we've got problems that mean we should
abort execution entirely. I suspect that checking this property in
ykllvm is fairly easy, so perhaps we'll soon be able to remove the
`FIXME` anyway!