Implement `yk_promote`
A call to `yk_promote` tells yk to treat a runtime value as a constant
in a trace. This makes future optimisations -- such as constant
propagation and function elision -- more effective. This is implemented
during the trace building phase, where promoted values are replaced with
a constant value and a guard instruction.

`yk_promote` is currently only implemented for word-size integers.