Add a "temporary" `RegConstraint`.
This informs the register allocator that we need a temporary/scratch
register. This allows us to use an arbitrary (hopefully free!) register
for calls rather than forcibly using r12.

Currently only the general purpose registers support `Temporary`: if we
find we need to we can later add support to floating point registers.