Get rid of the last use of `WR0`.
Floating point constants are a pain because they have to go through a
general purpose register. This commit moves from using r12, so we can
finally unmark that as "reserved" (i.e. normal code can now allocate
r12) to an arbitrary other register (`rax`, which is the register we are
most likely to clobber elsewhere). However, we still have to push/pop to
avoid clobbering that register, so we don't gain anything (yet) by
choosing one register over another.

This commit highlights a weakness in our register allocator API. We
might need to express constraints on both general and floating point
registers simultaneously. If we're lucky (and this is what I'm currently
leaning to) we might be able handle this internally to the register
allocator by storing the current `avoid` set. Fixing this properly is a
problem for another day, though.