Flip the lang_tests name-running strategy.
Before we only ran tests with "newcg.c" in their names: now we don't run
tests with "old.c" in their name.

The renaming strategy is (in essence):

  1. Remove a test T.c if there is a ported test T.newcg.c.
  2. For any unported test T.c, rename it to T.old.c.
  3. For a test T.newcg.c, rename it to T.c.

There are a few tests that have diverged between old and new codegens,
but none of the old versions strikes me as worth reconsidering.