Don't preemptively spill variables before a guard.
Previously the only way to pass an instruction's value to a guard was
via spill -- so if a value was referenced in a guard and not already
spilled, we had to spill it just before the guard, even if there was no
other reason to do so.

This commit removes this penalty from the fast path: guards and deopt
(sort of...) can take values from registers rather than spills.

For example here's a guard from a test before this commit:

```
 ; guard true, %304, [<lots of live variables>]
 : mov [rbp-0x930], r10
 : mov [rbp-0x938], r9
 : mov [rbp-0x940], r8
 : mov [rbp-0x948], rdx
 : mov [rbp-0x950], rcx
 : mov [rbp-0x958], rax
 : mov [rbp-0x960], r11
 : cmp r15b, 0x01
 : jnz 0xCB3
 ; %306: ptr = load %6
```

and after:

```
 ; guard true, %304, [<lots of live variables>]
 : cmp r15b, 0x01
 : jnz 0xC4A
 ; %306: ptr = load %6
```

This is a fairly extreme example (most guards tend to spill 1 or 2
variables), but it shows 7 spills being removed entirely.

Now to some extent, what this commit does is shuffle the problem around
a bit: we still have to pass lots of register values to deopt, and
there's not enough room to do so. We could use the SysV ABI in a clever
way, but instead we push all registers (including reserved registers...
which is silly... but easy) in a failing guard to the stack, and then
fish values out in deopt.

What this means is that the happy path (i.e. a guard that succeeds)
doesn't encounter any spills, but the unhappy path encounters lots.
Since the latter case is slow anyway, we don't really notice the costs:
for benchmarks which have lots of failing guards, this commit has no
real impact. For benchmarks which stay in JITed code we have a noticable
benefit (e.g. big_loop.lua speeds up by a bit over 7%).

There's more we could do to make this more sophisticated and complete
(e.g. we'll hit a `TODO` if/when we hit a guard which references a
floating point variable --- none of our tests does so), but this is a
step in the right direction.