Remove dynasm-x64-ness from outside the x64 backend.
As part of https://github.com/ykjit/yk/pull/1329, which fiddles with
trace inputs, it was easiest to put knowledge about dynasm and x64 into
various parts of the system beyond just the x64 backend. I thought that
removing this would be quite easy, but actually many more assumptions
than are first obvious were embedded as part of the PR.

This commit tries to remove all of those assumptions, but doing so isn't
always pretty.

The basic idea is to use yksmp's registers in trace_builder and the JIT
IR rather than backend (i.e. x64) registers. Amongst the challenges are
that I don't know what the fields (well, I can partly guess) but
(particularly) or values of those fields in yksmp are. For example we
have this code:

```rust
yksmp::Location::Register(3, ..) => {
    VarLocation::Register(reg_alloc::Register::GP(Rq::RBX))
}
```

What is the `3`? What about the other values? Is it safe to map things
in this way? I have tried to be conservative in each place where I don't
really understand why we're doing such things.

This creates a particular challenge for the JIT IR parser. Previously we
assigned every instruction to either `RBX` or `XMM0` but that isn't
going to work for very long, because optimisation passes could work out
they're aliases and optimise things away. This commit uses distinct
(general purpose? depends on yksmp's semantics!) registers for most
types but puts floating point types on the heap. This is particularly
messy because we can only assign trace inputs to 3 GP registers before
the backend borks. I don't know why.

Overall I'm not super happy with this commit, because it perpetuates a
number of magic constants that I don't understand. However, I'm
reasonably sure that I haven't made things much worse (I've been very
conservative), and it does break a dependency that needed breaking, and
that would get harder to break the longer we leave things. So, overall,
I think it's a win.

I'm hoping that Lukas will be able to cast his expert eye over this
later and improve some of this in a future PR.